Finished implementation of LoadGrottoAndWarpPlayer

+++ src/field_effect.c
@@ -3928,7 +3928,7 @@ static void Task_MoveDeoxysRock(u8 taskId)
 
 u32 GetCurrentMap(void)//Used for PSF Flying Blind
 {
-    u16 map = (gSaveBlock1Ptr->location.mapGroup << 8) | gSaveBlock1Ptr->location.mapNum;
+    u32 map = (gSaveBlock1Ptr->location.mapGroup << 8) | gSaveBlock1Ptr->location.mapNum;
 
     gSpecialVar_Result = map;
     return map;
